UserInformationService.java has now the method authorizeUserForWorkspace.
This method authorizes if a user is allowed to access a particular workspace. There are three overloads of this method with different arguments.

The response is null, if no workspace was found that meets the criteria. This means either the specified Workspace does not exist or the user has no access to it.

If the workspace exists and the user has access to it, then the workspace Entity is returned.