FileRestController.java and FileService.java have each three new methods, which work hand in hand.
Three endpoints have been added:

"/files/{workspaceId} is for getting all the folders and files which belong to a workspace, which is specified with the workspaceId via the PathVariable.

"/folder/{folderId}" is for getting all the children of a folder. The folder is specified via the PathVariable.

"/file/{fileId}" is for getting the information of a specific file, NOT a folder. The folder is specified via the PathVariable.